ports/esp32/settings_writer/requirements.txt
paho-mqtt==2.1.0
blinker==1.9.0
click==8.3.0
Flask==3.1.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
paho-mqtt==2.1.0
Werkzeug==3.1.3
aniso8601==10.0.1
blinker==1.9.0
click==8.3.0
Flask==3.1.2
Flask-RESTful==0.3.10
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
paho-mqtt==2.1.0
pytz==2025.2
six==1.17.0
Werkzeug==3.1.3
